Move contrib elementwise operators to a separate file
Move the `contrib` submodule of `src/ops/unary_elementwise.rs` to
`src/ops/unary_elementwise/contrib.rs` and move the tests for the
contrib operators into a `tests` module in that file.
